Refactor Linux desktop entry helpers into its own file
