+++ gui/src/renderer/app.tsx
@@ -20,7 +20,7 @@ import { ICurrentAppVersionInfo } from '../main';
 import { loadTranslations, messages, relayLocations } from '../shared/gettext';
 import { IGuiSettingsState, SYSTEM_PREFERRED_LOCALE_KEY } from '../shared/gui-settings-state';
 import { IpcRendererEventChannel, IRelayListPair } from '../shared/ipc-event-channel';
-import ISplitTunnelingApplication from '../shared/linux-split-tunneling-application';
+import { ILinuxSplitTunnelingApplication } from '../shared/application-types';
 import { getRendererLogFile, setupLogging } from '../shared/logging';
 import consumePromise from '../shared/promise';
 import History from './lib/history';
@@ -416,7 +416,7 @@ export default class AppRenderer {
     return IpcRendererEventChannel.splitTunneling.getApplications();
   }
 
-  public launchExcludedApplication(application: ISplitTunnelingApplication | string) {
+  public launchExcludedApplication(application: ILinuxSplitTunnelingApplication | string) {
     consumePromise(IpcRendererEventChannel.splitTunneling.launchApplication(application));
   }
