--- gui/src/shared/linux-split-tunneling-application.ts
@@ -1,9 +1,0 @@
-type Warning = 'launches-in-existing-process' | 'launches-elsewhere';
-
-export default interface ISplitTunnelingApplication {
-  absolutepath: string;
-  name: string;
-  exec: string;
-  icon?: string;
-  warning?: Warning;
-}
